core/pyproject.toml
[tool.poetry]
name = "fullstack-auth-posts"
version = "0.1.0"
description = ""
authors = ["radthenone <krzyzak.codes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.8"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"
django-cors-headers = "^4.0.0"
django-extensions = "^3.2.3"
pre-commit = "^3.3.2"
celery = "^5.3.0"
redis = "^4.5.5"
flower = "^1.2.0"
django-celery-beat = "^2.5.0"
django-filter = "^23.2"
watchfiles = "^0.19.0"
werkzeug = "^2.3.4"
pillow = "^10.1.0"
argon2-cffi = "^21.3.0"
django-redis = "^5.3.0"
django-log-request-id = "^2.1.0"
filetype = "1.2.0"
black = "23.7.0"
tox = "^4.11.4"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
django-stubs = "^4.2.1"
factory-boy = "^3.2.1"
faker = "^21.0.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"
coverage = "^7.2.7"
django-silk = "^5.0.3"
python-json-logger = "^2.0.7"
ruff = "^0.0.277"
pytest-factoryboy = "^2.5.1"
black = "^23.7.0"
pytest-cases = "^3.8.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = [
    "I", # isort
    "C",
    "E",
    "F",
    "W",
    "F",
    "B",
    "C4",
    "ARG",
    "SIM",
    "DJ",
]
src = ["core", "frontend"]
ignore = ["DJ01", 'W291', 'W292', 'W293']
unfixable = []
line-length = 88
extend-exclude = [
    ".git",
    "manage.py",
    "migrations",
    "static",
    "config",
    "docs",
    "node_modules",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"./manage.py" = ["INP001"]
"*/core/*" = ["E501"]

[tool.ruff.isort]
known-third-party = [
    "django",
    "rest-framework",
    "celery",
    "redis",
    "flower",
    "django-celery-beat",
    "django-filter",
    "django-extensions",
    "django-stubs",
    "factory-boy",
    "faker",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
addopts = "--cov=apps --cov-report lcov --cov-report term-missing --ds=config.settings.testing --reuse-db -p no:warnings --no-migrations"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py"
]

python_classes = [
    "Test",
]
testpaths = [
    "tests",
    "integration"
]
filterwarnings = [
    "ignore:.*U.*mode is deprecated:DeprecationWarning"
]

[tool.tox]
envlist = [
    "py311",
    "py310",
    "py39",
]